Solution for Day 9

+++ AdventOfCode.Tests/Day02Solver_Tests.cs
@@ -15,7 +15,7 @@ namespace AdventOfCode.Tests
         public void TestPart1Solution(string inputFile, int expected)
         {
             // Arrange
-            int[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
+            long[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
             if (inputFile.Contains("puzzle.input"))
             {
                 // Set position 1 and 2 values to IntCode computer error state 1202
@@ -35,7 +35,7 @@ namespace AdventOfCode.Tests
         public void TestPart2Solution(string inputFile, int expected)
         {
             // Arrange
-            int[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
+            long[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
 
             // Act
             var result = Day2Solver.Part2Solution(input);
