+++ AdventOfCode.Tests/Day07Solver_Tests.cs
@@ -13,7 +13,7 @@ namespace AdventOfCode.Tests
         public void TestPart1Solution(string inputFile, int expected)
         {
             // Arrange
-            int[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
+            long[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
 
             // Act
             var result = Day7Solver.Part1Solution(input);
@@ -29,7 +29,7 @@ namespace AdventOfCode.Tests
         public void TestPart2Solution(string inputFile, int expected)
         {
             // Arrange
-            int[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
+            long[] input = Helpers.ReadIntCodeInput($"../../../{inputFile}");
 
             // Act
             var result = Day7Solver.Part2Solution(input);
