+++ Day 5 Solver/Day5Solver.cs
@@ -6,14 +6,14 @@ namespace Day_5_Solver
 {
     public static class Day5Solver
     {
-        public static int Part1Solution(int[] input)
+        public static long Part1Solution(long[] input)
         {
             var program = new IntCodeProgram(input, 1);
             program.Run();
             return program.Output.Last();
         }
 
-        public static int Part2Solution(int[] input)
+        public static long Part2Solution(long[] input)
         {
             var program = new IntCodeProgram(input, 5);
             program.Run();
